cleanup per review
